+++ tests/test_queue.py
@@ -319,12 +319,21 @@ class TestQueue(RQTestCase):
         self.assertEqual(job.meta['foo'], 'bar')
         self.assertEqual(job.meta['baz'], 42)
 
+    def test_job_timeout(self):
+        """Timeout can be passed via job_timeout argument"""
+        queue = Queue()
+        job = queue.enqueue(echo, 1, timeout=15)
+        self.assertEqual(job.timeout, 15)
+
+        job = queue.enqueue(echo, 1, job_timeout=15)
+        self.assertEqual(job.timeout, 15)
+
     def test_enqueue_explicit_args(self):
         """enqueue() works for both implicit/explicit args."""
         q = Queue()
 
         # Implicit args/kwargs mode
-        job = q.enqueue(echo, 1, timeout=1, result_ttl=1, bar='baz')
+        job = q.enqueue(echo, 1, job_timeout=1, result_ttl=1, bar='baz')
         self.assertEqual(job.timeout, 1)
         self.assertEqual(job.result_ttl, 1)
         self.assertEqual(
@@ -337,7 +346,7 @@ class TestQueue(RQTestCase):
             'timeout': 1,
             'result_ttl': 1,
         }
-        job = q.enqueue(echo, timeout=2, result_ttl=2, args=[1], kwargs=kwargs)
+        job = q.enqueue(echo, job_timeout=2, result_ttl=2, args=[1], kwargs=kwargs)
         self.assertEqual(job.timeout, 2)
         self.assertEqual(job.result_ttl, 2)
         self.assertEqual(
@@ -645,11 +654,18 @@ class TestFailedQueue(RQTestCase):
         self.assertEqual(int(job_from_queue.result_ttl), 10)
 
     def test_async_false(self):
-        """Job executes and cleaned up immediately if async=False."""
-        q = Queue(async=False)
+        """Job executes and cleaned up immediately if is_async=False."""
+        q = Queue(is_async=False)
         job = q.enqueue(some_calculation, args=(2, 3))
         self.assertEqual(job.return_value, 6)
         self.assertNotEqual(self.testconn.ttl(job.key), -1)
+
+    def test_is_async(self):
+        """Queue exposes is_async as a property."""
+        inline_queue = Queue(is_async=False)
+        self.assertFalse(inline_queue.is_async)
+        async_queue = Queue(is_async=True)
+        self.assertTrue(async_queue.is_async)
 
     def test_custom_job_class(self):
         """Ensure custom job class assignment works as expected."""